+++ framework-generators/direct-adapter-core/src/test/java/uk/gov/justice/services/adapter/direct/DirectAdapterIT.java
@@ -22,6 +22,7 @@ import uk.gov.justice.services.common.converter.StringToJsonObjectConverter;
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
 import uk.gov.justice.services.core.dispatcher.DispatcherCache;
+import uk.gov.justice.services.core.dispatcher.DispatcherConfiguration;
 import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
 import uk.gov.justice.services.core.dispatcher.EmptySystemUserProvider;
 import uk.gov.justice.services.core.dispatcher.EnvelopePayloadTypeConverter;
@@ -165,7 +166,9 @@ public class DirectAdapterIT {
             ComponentNameExtractor.class,
             JndiAppNameProvider.class,
 
-            SchemaValidationErrorMessageGenerator.class
+            SchemaValidationErrorMessageGenerator.class,
+
+            DispatcherConfiguration.class
     })
     public WebApp war() {
         return new WebApp()