+++ framework-generators/messaging-adapter-generator/src/test/java/uk/gov/justice/subscription/jms/it/JmsEndpointGenerationIT.java
@@ -29,6 +29,7 @@ import uk.gov.justice.services.core.accesscontrol.AllowAllPolicyEvaluator;
 import uk.gov.justice.services.core.accesscontrol.DefaultAccessControlService;
 import uk.gov.justice.services.core.accesscontrol.PolicyEvaluator;
 import uk.gov.justice.services.core.dispatcher.DispatcherCache;
+import uk.gov.justice.services.core.dispatcher.DispatcherConfiguration;
 import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
 import uk.gov.justice.services.core.dispatcher.EmptySystemUserProvider;
 import uk.gov.justice.services.core.dispatcher.EnvelopePayloadTypeConverter;
@@ -180,7 +181,9 @@ public class JmsEndpointGenerationIT extends AbstractJmsAdapterGenerationIT {
             JmsLoggerMetadataAdder.class,
             ComponentNameExtractor.class,
 
-            JndiAppNameProvider.class
+            JndiAppNameProvider.class,
+
+            DispatcherConfiguration.class
     })
     public WebApp war() {
         return new WebApp()