+++ metrics/metrics-interceptor/src/test/java/uk/gov/justice/services/metrics/interceptor/it/MetricsIT.java
@@ -27,6 +27,7 @@ import uk.gov.justice.services.core.annotation.FrameworkComponent;
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
 import uk.gov.justice.services.core.dispatcher.DispatcherCache;
+import uk.gov.justice.services.core.dispatcher.DispatcherConfiguration;
 import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
 import uk.gov.justice.services.core.dispatcher.EmptySystemUserProvider;
 import uk.gov.justice.services.core.dispatcher.EnvelopePayloadTypeConverter;
@@ -142,7 +143,9 @@ public class MetricsIT {
             GlobalValueProducer.class,
             JsonSchemaLoader.class,
             JsonEnvelopeProvider.class,
-            SchemaCatalogResolverProducer.class
+            SchemaCatalogResolverProducer.class,
+
+            DispatcherConfiguration.class
     })
     public WebApp war() {
         return new WebApp()